+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### New Components
 - ScriptInput that uses a script to generate a list of Items
-- BulkFilter a new type of filter that does pre-processing with the list of Items for filters that need to handle all of the Items at once for efficiency purposes.
+- BulkFilter a new type of filter that does pre-processing with the list of Items for filters that need to handle all of the Items at once for efficiency purposes
 - ScriptFilter that uses a script to validate items
+- OpenAITransformer that uses OpenAI LLM completition to replace file content
 
 ### Bugs
 - Fix ScriptCommand to return when run_on_changes is set to true for batch runs and no changes are present.